cleaned up pageSchema and added collection field array functionality
